Include spell name in cast events and centralize status effects
