+++ client/src/components/Zombies/attributes/PlayerTurnActions.js
@@ -9,7 +9,6 @@ import { Button, Modal, Card, Table } from "react-bootstrap";
 import UpcastModal from './UpcastModal';
 import sword from "../../../images/sword.png";
 import proficiencyBonus from '../../../utils/proficiencyBonus';
-import hasteIcon from "../../../images/spell-haste-icon.png";
 
 // Dice rolling helper used by calculateDamage and component actions
 function rollDice(numberOfDiceValue, sidesOfDiceValue) {
@@ -90,7 +89,6 @@ const PlayerTurnActions = React.forwardRef(
       availableSlots = { regular: {}, warlock: {} },
       longRestCount = 0,
       shortRestCount = 0,
-      onEffectsChange,
     },
     ref
   ) => {
@@ -150,36 +148,6 @@ const [isFumble, setIsFumble] = useState(false);
 const [showUpcast, setShowUpcast] = useState(false);
 const [pendingSpell, setPendingSpell] = useState(null);
 
-// Track active status effects
-const [activeEffects, setActiveEffects] = useState([]);
-const removeEffect = (name) =>
-  setActiveEffects((prev) => prev.filter((e) => e.name !== name));
-
-useEffect(() => {
-  onEffectsChange?.(activeEffects);
-}, [activeEffects, onEffectsChange]);
-
-useEffect(() => {
-  const handlePass = () => {
-    setActiveEffects((prev) =>
-      prev
-        .map((e) =>
-          e.name === 'Haste'
-            ? { ...e, remaining: (e.remaining || 0) - 1 }
-            : e
-        )
-        .filter((e) => e.name !== 'Haste' || e.remaining > 0)
-    );
-  };
-  window.addEventListener('pass-turn', handlePass);
-  return () => window.removeEventListener('pass-turn', handlePass);
-}, []);
-
-useEffect(() => {
-  // Clear effects on rest
-  setActiveEffects([]);
-}, [longRestCount, shortRestCount]);
-
   const applyUpcast = (spell, level, crit, slotType) => {
     const diff = level - (spell.level || 0);
     let extra;
@@ -207,7 +175,7 @@ useEffect(() => {
     );
     if (value === null) return;
     updateDamageValueWithAnimation(value);
-    onCastSpell?.({ level, slotType, castingTime: spell.castingTime });
+    onCastSpell?.({ level, slotType, castingTime: spell.castingTime, name: spell.name });
   };
 
   const handleSpellsButtonClick = (spell, crit = false) => {
@@ -218,9 +186,6 @@ useEffect(() => {
       return;
     }
     applyUpcast(spell, spell.level, crit || isCritical);
-    if (spell.name === 'Haste') {
-      setActiveEffects((prev) => [...prev, { name: 'Haste', icon: hasteIcon, remaining: 10 }]);
-    }
   };
 
 const handleDamageClick = () => {
