+++ client/src/components/Zombies/attributes/PlayerTurnActions.test.js
@@ -168,6 +168,7 @@ describe('PlayerTurnActions spell casting', () => {
       level: spell.level,
       slotType: undefined,
       castingTime: spell.castingTime,
+      name: spell.name,
     });
   });
